Add tests for exceding max allowed size.
Also fix the max int to an int32.  When the code was originally written an
int was always int32, however that is no longer the case in newer version
of Go.

While here add an additional check for the padded size in
DecodeFixedOpaque exceeding the max allowed size which was exposed by the
new tests.