Expose Decode/Encode functions on Decoder/Encoder.
This commit exposes two new functions which make it a little more
intuitive to mix automatic and manual encoding and decoding as well as
make the API more consistent with other packages in the standard library,
that perform a similar function, such as gob and json.

The first is Decoder.Decode and it operates identically to Unmarshal with
the exception of operating on the reader associated with the Decoder.

The second is Encoder.Encode and it operates identically to Marhsal with
the exception of operating on the writer associated with the Encoder.

This was proposed by @marcopeereboom.