Remove xdr_test package to ease forking.
The usage of absolute import paths in testsuite is not compatible
with allowing a package to easily fork. One has to replace all
global paths in test files. I tried to use relative import paths
such as "../xdr2" but there appears to be a Go 1.5 bug for which
the internal_test.go functions TstEncode/TstDecode are not exposed
to tests.